requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2

# AI/ML Stack (Essential only)
langchain==0.1.0
langchain-openai==0.0.5
langchain-huggingface==0.0.1
langchain-pinecone==0.0.3
openai==1.12.0
pinecone-client==3.0.0
httpx==0.25.2
sentence-transformers==2.2.2

# Document Processing (Essential only)
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# Utilities
aiofiles==23.2.1